python/enact-protocol/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enact-protocol"
version = "0.1.0"
description = "Python SDK for ENACT Protocol — trustless escrow for AI agents on TON"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "ENACT Protocol" }]
keywords = ["ton", "escrow", "ai-agents", "blockchain", "enact", "jobs", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
]
dependencies = [
    "tonutils>=0.4.0,<1.0",
    "pytoniq-core>=0.1.40,<1.0",
    "pynacl>=1.5.0,<2.0",
    "httpx>=0.27,<1.0",
    "pydantic>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[project.urls]
Homepage = "https://enact.info"
Documentation = "https://enact.info/docs/python-sdk"
Repository = "https://github.com/ENACT-protocol/enact-protocol"
Issues = "https://github.com/ENACT-protocol/enact-protocol/issues"

[tool.hatch.build.targets.wheel]
packages = ["enact_protocol"]

[tool.hatch.build.targets.sdist]
include = [
    "enact_protocol",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
